+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `@homebridge-plugins/homebridge-meross` will be documented in this file.
 
-## v10.19.1 (Pending Release)
+## v10.19.1 (2026-08-09)
 
 ### Changed